line_bot_DJI/requirements.txt
line-bot-sdk
flask
beautifulsoup4


